requirements.txt
python-dotenv
cryptography
colorama